# openapi.yaml
openapi: 3.1.0
info:
  title: GABAI API
  version: 0.1.0
  description: |
    Citizen service request classification and routing for a barangay LGU.

    Residents submit free-text requests; a classifier assigns a category and an
    urgency level, each with its own confidence score. A request is auto-routed
    only when `min(category_confidence, urgency_confidence)` meets the configured
    threshold; otherwise it enters the manual review queue for a human to label.

    **Conventions**

    - JSON bodies use `snake_case`.
    - Timestamps are ISO-8601 with a UTC offset (`2026-08-30T04:15:00Z`).
    - List endpoints use offset pagination and return a `PaginatedResponse`.
    - Errors use a single envelope; see `Error`.
    - Authorization is enforced server-side on every endpoint. A client hiding a
      control is not access control.
    - Categories are **data, not an enum**. Never hardcode them; read
      `GET /api/categories`.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: /
    description: Same origin as the deployed frontend
...
security:
  - cookieAuth: []
paths:
...
  /api/requests:
    post:
      tags: [requests]
      summary: Submit a request
      description: |
        Persists the request and returns it immediately with status `submitted`
        and a reference number. Classification runs after the response, so the
        citizen never waits on inference; poll the detail endpoint (or watch
        notifications) for the classified result.

        Only authenticated citizens may submit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestCreate"
      responses:
        "201":
          description: Accepted and persisted. Not yet classified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [requests]
      summary: List requests, scoped to the caller's role
      description: |
        Scope is applied server-side and cannot be widened by query parameters:

        - **citizen**, only their own requests
        - **staff**, requests assigned to them, plus requests in a category they
          hold an active routing rule for, plus everything in `under_review`. The
          union matters twice: deactivating a rule must not strip a staff member
          of requests already assigned to them, and a request awaiting manual
          classification cannot be scoped by a category that does not exist yet.
        - **admin**, all requests

        The `under_review` clause is a deliberate widening of least privilege.
        Unclassified requests are visible to all staff because the category that
        would otherwise scope them is exactly what is in doubt; every review
        decision is attributed in the audit log.

        Default ordering is role-dependent and chosen by the server, not the
        client. Staff and admin get highest urgency first, then oldest first,
        which is the order the queue is worked in. Citizens get newest first, because
        a resident opening their list is looking for the request they just
        filed, not the most urgent one.
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: status
          in: query
          description: Repeat the parameter to match any of several statuses.
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RequestStatus"
        - name: category_id
          in: query
          schema:
            type: integer
        - name: urgency
          in: query
          schema:
            $ref: "#/components/schemas/Urgency"
        - name: q
          in: query
          description: Case-insensitive substring match on description and reference number.
          schema:
            type: string
            maxLength: 200
        - name: sort
          in: query
          schema:
            type: string
            enum: [urgency_desc, created_asc, created_desc, updated_desc]
            default: urgency_desc
      responses:
        "200":
          description: A page of requests.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/RequestSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/requests/{request_id}:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    get:
      tags: [requests]
      summary: Request detail
      description: |
        Same role scoping as `GET /api/requests`, including the `under_review`
        clause. Otherwise every row in the review queue would 404 on click.

        Returns 404, not 403, when the request exists but is outside the caller's
        scope. A 403 would confirm the id is real and let a staff member map
        requests outside their categories by probing ids.
      responses:
        "200":
          description: The request, with its status history and attachments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/requests/{request_id}/status:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    patch:
      tags: [requests]
      summary: Change status
      description: |
        Staff and admin only. Writes a `status_history` row and an audit entry,
        and notifies the citizen. Illegal transitions are rejected with 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Not a legal transition from the current status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/requests/{request_id}/classification:
    parameters:
      - $ref: "#/components/parameters/RequestId"
    patch:
      tags: [review]
      summary: Set the final category and urgency
      description: |
        Staff and admin. Sets `final_category_id` and `final_urgency`, then
        re-runs routing.

        Works on **any** request the caller can see, not only items in the review
        queue, because the case where correction matters most is when the model
        was confident and wrong, and that request never reaches the queue. The
        model's original prediction is never overwritten, so every correction
        stays measurable against it.

        **Re-routing can hand the request to someone else.** Changing the
        category re-runs the routing rules, so a request may be reassigned to a
        different staff member, after which the correcting staff member no
        longer sees it, because it is no longer theirs. That is intended: they
        handed it off. The citizen is notified of the reassignment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewDecision"
      responses:
        "200":
          description: |
            Labelled and routed. `assigned_staff` may differ from before the call.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Request is `resolved` or `closed`. A finished request is not
            relabelled. Re-routing one would resurrect it into an active queue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: |
            The body failed validation, or `final_category_id` is not a category
            that exists and is switched on.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------- categories ----
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gabai_session
      description: |
        HttpOnly session cookie set by `POST /api/auth/login`. Browsers attach it
        automatically; send credentialed requests (`credentials: "include"` on
        cross-origin fetches).
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: No session, or the session has expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated, but not allowed. The role does not permit this action,
        the account is still waiting for approval, it is on a temporary
        password and has to change it first, or it hasn't accepted the current
        Terms of Use and Privacy Notice.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        No such resource, or it exists outside the caller's scope. The two are
        deliberately indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: The body failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Safe to branch on.
              examples: [not_found, forbidden, validation_error, invalid_transition]
            message:
              type: string
              description: Human-readable, safe to show a barangay staff member.
            details:
              type: array
              description: Per-field problems, present on validation errors.
              items:
                type: object
                required: [field, message]
                properties:
                  field:
                    type: string
                    examples: [description]
                  message:
                    type: string
                    examples: ["Field is required."]
    PaginatedResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          description: Total matching rows, ignoring limit and offset.
        limit:
          type: integer
        offset:
          type: integer
...
    Role:
      type: string
      enum: [citizen, staff, admin]
    Urgency:
      type: string
      enum: [low, medium, high]
      description: A learned model output, not derived from keywords or rules.
    RequestStatus:
      type: string
      enum:
        - submitted
        - classified
        - routed
        - under_review
        - in_progress
        - resolved
        - closed
      description: |
        `submitted` on creation. Then either `classified` → `routed` when both
        confidence scores clear the threshold, or `under_review` when either does
        not. From `routed`: `in_progress` → `resolved` → `closed`.

    # --------------------------------------------------------- users ------
...
    UserSummary:
      type: object
      description: Identity only. Used where a full user record would leak PII.
      required: [id, full_name, role]
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
...
    Category:
      type: object
      required: [id, slug, name, description, is_active]
      properties:
        id:
          type: integer
        slug:
          type: string
          description: |
            Stable machine identifier, and the exact label string used in the
            training dataset. Changing it desynchronises the model from the
            database.
          examples: [road_infrastructure]
        name:
          type: string
          description: Display label.
        description:
          type: [string, "null"]
          description: What belongs in this category. Shown as help text.
        is_active:
          type: boolean
...
    RequestCreate:
      type: object
      required: [description]
      properties:
        description:
          type: string
          minLength: 10
          maxLength: 5000
          description: Free text in Filipino, English, or a mix.
    RequestSummary:
      type: object
      description: List rows. Enough for the staff queue without loading history.
      required:
        - id
        - reference_number
        - description
        - category
        - urgency
        - category_confidence
        - urgency_confidence
        - status
        - assigned_staff
        - created_at
      properties:
        id:
          type: integer
        reference_number:
          type: string
          examples: [GAB-2026-00042]
        description:
          type: string
        category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
          description: Effective category. The human label if set, otherwise the prediction.
        urgency:
          oneOf:
            - $ref: "#/components/schemas/Urgency"
            - type: "null"
          description: Effective urgency, same rule as `category`.
        category_confidence:
          type: [number, "null"]
          format: float
          minimum: 0
          maximum: 1
        urgency_confidence:
          type: [number, "null"]
          format: float
          minimum: 0
          maximum: 1
          description: |
            Carried on list rows so the staff queue can show confidence without a
            detail fetch per row. Surfacing the score makes the classifier legible
            to the operator rather than opaque.
        status:
          $ref: "#/components/schemas/RequestStatus"
        assigned_staff:
          oneOf:
            - $ref: "#/components/schemas/UserSummary"
            - type: "null"
        created_at:
          type: string
          format: date-time
    Request:
      type: object
      description: |
        Nullable fields are always present in the response, carrying `null` when
        unset. They are listed as required so a client cannot confuse "absent"
        with "not yet classified".
      required:
        - id
        - reference_number
        - citizen
        - description
        - predicted_category
        - predicted_urgency
        - category_confidence
        - urgency_confidence
        - final_category
        - final_urgency
        - category
        - urgency
        - status
        - assigned_staff
        - model_version
        - classified_at
        - created_at
        - updated_at
        - resolved_at
        - attachments
        - status_history
      properties:
        id:
          type: integer
        reference_number:
          type: string
          description: |
            Human-readable handle a resident can quote in person or by phone.
            Drawn from a database sequence, so gaps after a rolled-back
            transaction are expected and not a defect.
          examples: [GAB-2026-00042]
        citizen:
          $ref: "#/components/schemas/UserSummary"
        description:
          type: string
        predicted_category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
          description: Null until classification runs.
        predicted_urgency:
          oneOf:
            - $ref: "#/components/schemas/Urgency"
            - type: "null"
        category_confidence:
          type: [number, "null"]
          format: float
          minimum: 0
          maximum: 1
        urgency_confidence:
          type: [number, "null"]
          format: float
          minimum: 0
          maximum: 1
          description: |
            Scored independently of the category head. A request is auto-routed
            only if `min(category_confidence, urgency_confidence)` meets the
            threshold, one gate over both labels, so a request is never routed
            on a label the model was unsure of.
        final_category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
          description: |
            Stays null unless a human sets it. Null therefore means "no human has
            labelled this", which is distinct from "a human agreed with the
            model". Collapsing the two would erase the correction record.
        final_urgency:
          oneOf:
            - $ref: "#/components/schemas/Urgency"
            - type: "null"
        category:
          oneOf:
            - $ref: "#/components/schemas/Category"
            - type: "null"
          description: Effective value. `final_category` if set, otherwise `predicted_category`.
        urgency:
          oneOf:
            - $ref: "#/components/schemas/Urgency"
            - type: "null"
          description: Effective value, same rule as `category`.
        status:
          $ref: "#/components/schemas/RequestStatus"
        assigned_staff:
          oneOf:
            - $ref: "#/components/schemas/UserSummary"
            - type: "null"
        model_version:
          type: [string, "null"]
          description: |
            Identifier of the exported model that produced the prediction, matching
            the run name recorded during training. Without it, predictions made
            before and after a retrain are indistinguishable.
        classified_at:
          type: [string, "null"]
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type: [string, "null"]
          format: date-time
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/Attachment"
        status_history:
          type: array
          description: Oldest first. Drives the citizen-facing status timeline.
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
    StatusUpdate:
      type: object
      required: [to_status]
      properties:
        to_status:
          type: string
          enum: [in_progress, resolved, closed]
          description: |
            Deliberately narrower than `RequestStatus`. The earlier statuses are
            set by the system during classification and routing, so a staff UI
            should not be able to offer them. The illegal option is unrepresentable
            in the client rather than rejected by the server at runtime.
        note:
          type: string
          maxLength: 2000
          description: Resolution note or reason. Shown to the citizen.
    StatusHistoryEntry:
      type: object
      required: [id, from_status, to_status, actor, note, created_at]
      properties:
        id:
          type: integer
        from_status:
          oneOf:
            - $ref: "#/components/schemas/RequestStatus"
            - type: "null"
          description: Null on the first entry, which has no predecessor.
        to_status:
          $ref: "#/components/schemas/RequestStatus"
        actor:
          oneOf:
            - $ref: "#/components/schemas/UserSummary"
            - type: "null"
          description: Null when the system made the transition, e.g. auto-routing.
        note:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
    ReviewDecision:
      type: object
      required: [final_category_id, final_urgency]
      properties:
        final_category_id:
          type: integer
        final_urgency:
          $ref: "#/components/schemas/Urgency"
        note:
          type: string
          maxLength: 2000
    Attachment:
      type: object
      required: [id, filename, mime_type, size_bytes, uploaded_at]
      properties:
        id:
          type: integer
        filename:
          type: string
          description: Original name, sanitised. Never used as the storage path.
        mime_type:
          type: string
        size_bytes:
          type: integer
        uploaded_at:
          type: string
          format: date-time

    # --------------------------------------------- routing and audit ------
...